pyproject.toml
[project]
name = "portfolio-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mosek>=11.0.11",
    "riskfolio-lib>=7.0.0",
    "yfinance>=0.2.54",
]
